fix PML after cleanup

+++ src/ElectroMagnSolver/PML_Solver2D_Bouchard.cpp
@@ -175,7 +175,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             sigma_x_p[i] = 0. ;
         }
         // Params for other cells (PML Media) when i>=3
-        for( unsigned int i = startpml; i<nx_p ; i++ ) {
+        for( int i = startpml; i< (int) nx_p ; i++ ) {
             //kappa_x_p[i] = 1. + (kappa_x_max - 1.) * pow( (i-startpml)*dx , kappa_power_pml_x ) / pow( length_x_pml , kappa_power_pml_x ) ;
             //sigma_x_p[i] = sigma_x_max * pow( (i-startpml)*dx , sigma_power_pml_x ) / pow( length_x_pml , sigma_power_pml_x ) ;
             kappa_x_p[i] = pml_kappa_[0]->valueAt((i-startpml)*dx/length_x_pml);
@@ -200,7 +200,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             sigma_x_d[i] = 0. ;
         }
         // Params for other cells (PML Media) when j>=4
-        for( unsigned int i = startpml+1 ; i<nx_d ; i++ ) {
+        for( int i = startpml+1 ; i< (int) nx_d ; i++ ) {
             //kappa_x_d[i] = 1. + (kappa_x_max - 1.) * pow( (i-startpml-0.5)*dx , kappa_power_pml_x ) / pow( length_x_pml , kappa_power_pml_x ) ;
             //sigma_x_d[i] = sigma_x_max * pow( (i-startpml-0.5)*dx , sigma_power_pml_x ) / pow( length_x_pml , sigma_power_pml_x ) ;
             kappa_x_d[i] = pml_kappa_[0]->valueAt((i-startpml-0.5)*dx/length_x_pml);
@@ -239,7 +239,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             }
         }
         if (ncells_pml_max[0] != 0 ){
-            for( unsigned int i = (nx_p-1)-(ncells_pml_max[0]-1) ; i<nx_p ; i++ ) {
+            for( int i = (nx_p-1)-(ncells_pml_max[0]-1) ; i< (int) nx_p ; i++ ) {
                 //kappa_x_p[i] = 1. + (kappa_x_max - 1.) * pow( ( i - ( (nx_p-1)-(ncells_pml_max[0]-1) ) )*dx , kappa_power_pml_x ) / pow( length_x_pml_xmax , kappa_power_pml_x ) ;
                 //sigma_x_p[i] = sigma_x_max * pow( (i - ( (nx_p-1)-(ncells_pml_max[0]-1) ) )*dx , sigma_power_pml_x ) / pow( length_x_pml_xmax , sigma_power_pml_x ) ;
                 kappa_x_p[i] = pml_kappa_[0]->valueAt((i - nx_p  + ncells_pml_max[0])*dx/length_x_pml_xmax);
@@ -254,7 +254,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             sigma_y_p[j] = 0. ;
         }
         // Params for other cells (PML Media) when j>=3
-        for( unsigned int j = startpml ; j<ny_p ; j++ ) {
+        for( int j = startpml ; j< (int) ny_p ; j++ ) {
             //kappa_y_p[j] = 1. + (kappa_y_max - 1.) * pow( (j-startpml)*dy , kappa_power_pml_y ) / pow( length_y_pml , kappa_power_pml_y ) ;
             //sigma_y_p[j] = sigma_y_max * pow( (j-startpml)*dy , sigma_power_pml_y ) / pow( length_y_pml , sigma_power_pml_y ) ;
             kappa_y_p[j] = pml_kappa_[1]->valueAt((j-startpml)*dy/length_y_pml);
@@ -280,7 +280,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             }
         }
         if (ncells_pml_max[0] != 0 ){
-            for( unsigned int i = (nx_p-1)-(ncells_pml_max[0]-1)+1 ; i<nx_d ; i++ ) {
+            for( int i = (nx_p-1)-(ncells_pml_max[0]-1)+1 ; i< (int) nx_d ; i++ ) {
                 //kappa_x_d[i] = 1. + (kappa_x_max - 1.) * pow( (i - ( (nx_p-1)-(ncells_pml_max[0]-1) ) - 0.5 )*dx , kappa_power_pml_x ) / pow( length_x_pml_xmax , kappa_power_pml_x ) ;
                 //sigma_x_d[i] = sigma_x_max * pow( (i - ( (nx_p-1)-(ncells_pml_max[0]-1) ) - 0.5 )*dx , sigma_power_pml_x ) / pow( length_x_pml_xmax , sigma_power_pml_x ) ;
                 kappa_x_d[i] = pml_kappa_[0]->valueAt((i - nx_p + ncells_pml_max[0] - 0.5 )*dx/length_x_pml_xmax);
@@ -295,7 +295,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             sigma_y_d[j] = 0. ;
         }
         // Params for other cells (PML Media) when j>=4
-        for( unsigned int j = startpml+1 ; j<ny_d ; j++ ) {
+        for( int j = startpml+1 ; j< (int) ny_d ; j++ ) {
             //kappa_y_d[j] = 1. + (kappa_y_max - 1.) * pow( (j-startpml-0.5)*dy , kappa_power_pml_y ) / pow( length_y_pml , kappa_power_pml_y ) ;
             //sigma_y_d[j] = sigma_y_max * pow( (j-startpml-0.5)*dy , sigma_power_pml_y ) / pow( length_y_pml , sigma_power_pml_y ) ;
             kappa_y_d[j] = pml_kappa_[1]->valueAt((j-startpml-0.5)*dy/length_y_pml);
@@ -309,7 +309,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
     }
 
     if ((min_or_max==0)&&(iDim==0)){
-        for( unsigned int i = 0 ; i<nx_p ; i++ ) {
+        for( int i = 0 ; i< (int) nx_p ; i++ ) {
             c1_p_zfield[i] = ( 2.*kappa_x_p[(nx_p-1)-i] - dt*sigma_x_p[(nx_p-1)-i] ) / ( 2.*kappa_x_p[(nx_p-1)-i] + dt*sigma_x_p[(nx_p-1)-i] ) ;
             c2_p_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_p[(nx_p-1)-i] + dt*sigma_x_p[(nx_p-1)-i] ) ;
             c3_p_yfield[i] = ( 2.*kappa_x_p[(nx_p-1)-i] - dt*sigma_x_p[(nx_p-1)-i] ) / ( 2.*kappa_x_p[(nx_p-1)-i] + dt*sigma_x_p[(nx_p-1)-i] ) ;
@@ -318,7 +318,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             c6_p_xfield[i] = ( 2.*kappa_x_p[(nx_p-1)-i] - dt*sigma_x_p[(nx_p-1)-i] ) ;
         }
 
-        for( unsigned int i = 0 ; i<nx_d ; i++ ) {
+        for( int i = 0 ; i< (int) nx_d ; i++ ) {
             c1_d_zfield[i] = ( 2.*kappa_x_d[(nx_d-1)-i] - dt*sigma_x_d[(nx_d-1)-i] ) / ( 2.*kappa_x_d[(nx_d-1)-i] + dt*sigma_x_d[(nx_d-1)-i] ) ;
             c2_d_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_d[(nx_d-1)-i] + dt*sigma_x_d[(nx_d-1)-i] ) ;
             c3_d_yfield[i] = ( 2.*kappa_x_d[(nx_d-1)-i] - dt*sigma_x_d[(nx_d-1)-i] ) / ( 2.*kappa_x_d[(nx_d-1)-i] + dt*sigma_x_d[(nx_d-1)-i] ) ;
@@ -328,7 +328,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
         }
     }
     else {
-        for( unsigned int i = 0 ; i<nx_p ; i++ ) {
+        for( int i = 0 ; i< (int) nx_p ; i++ ) {
             c1_p_zfield[i] = ( 2.*kappa_x_p[i] - dt*sigma_x_p[i] ) / ( 2.*kappa_x_p[i] + dt*sigma_x_p[i] ) ;
             c2_p_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_p[i] + dt*sigma_x_p[i] ) ;
             c3_p_yfield[i] = ( 2.*kappa_x_p[i] - dt*sigma_x_p[i] ) / ( 2.*kappa_x_p[i] + dt*sigma_x_p[i] ) ;
@@ -337,7 +337,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             c6_p_xfield[i] = ( 2.*kappa_x_p[i] - dt*sigma_x_p[i] ) ;
         }
 
-        for( unsigned int i = 0 ; i<nx_d ; i++ ) {
+        for( int i = 0 ; i< (int) nx_d ; i++ ) {
             c1_d_zfield[i] = ( 2.*kappa_x_d[i] - dt*sigma_x_d[i] ) / ( 2.*kappa_x_d[i] + dt*sigma_x_d[i] ) ;
             c2_d_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_d[i] + dt*sigma_x_d[i] ) ;
             c3_d_yfield[i] = ( 2.*kappa_x_d[i] - dt*sigma_x_d[i] ) / ( 2.*kappa_x_d[i] + dt*sigma_x_d[i] ) ;
@@ -348,7 +348,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
     } // End X
 
     if (min_or_max==0){
-        for( unsigned int j = 0 ; j<ny_p ; j++ ) {
+        for( int j = 0 ; j< (int) ny_p ; j++ ) {
             c1_p_xfield[j] = ( 2.*kappa_y_p[(ny_p-1)-j] - dt*sigma_y_p[(ny_p-1)-j] ) / ( 2.*kappa_y_p[(ny_p-1)-j] + dt*sigma_y_p[(ny_p-1)-j] ) ;
             c2_p_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_p[(ny_p-1)-j] + dt*sigma_y_p[(ny_p-1)-j] ) ;
             c3_p_zfield[j] = ( 2.*kappa_y_p[(ny_p-1)-j] - dt*sigma_y_p[(ny_p-1)-j] ) / ( 2.*kappa_y_p[(ny_p-1)-j] + dt*sigma_y_p[(ny_p-1)-j] ) ;
@@ -357,7 +357,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             c6_p_yfield[j] = ( 2.*kappa_y_p[(ny_p-1)-j] - dt*sigma_y_p[(ny_p-1)-j] ) ;
         }
 
-        for( unsigned int j = 0 ; j<ny_d ; j++ ) {
+        for( int j = 0 ; j< (int) ny_d ; j++ ) {
             c1_d_xfield[j] = ( 2.*kappa_y_d[(ny_d-1)-j] - dt*sigma_y_d[(ny_d-1)-j] ) / ( 2.*kappa_y_d[(ny_d-1)-j] + dt*sigma_y_d[(ny_d-1)-j] ) ;
             c2_d_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_d[(ny_d-1)-j] + dt*sigma_y_d[(ny_d-1)-j] ) ;
             c3_d_zfield[j] = ( 2.*kappa_y_d[(ny_d-1)-j] - dt*sigma_y_d[(ny_d-1)-j] ) / ( 2.*kappa_y_d[(ny_d-1)-j] + dt*sigma_y_d[(ny_d-1)-j] ) ;
@@ -367,7 +367,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
         }
     }
     else if (min_or_max==1){
-        for( unsigned int j = 0 ; j<ny_p ; j++ ) {
+        for( int j = 0 ; j< (int) ny_p ; j++ ) {
             c1_p_xfield[j] = ( 2.*kappa_y_p[j] - dt*sigma_y_p[j] ) / ( 2.*kappa_y_p[j] + dt*sigma_y_p[j] ) ;
             c2_p_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_p[j] + dt*sigma_y_p[j] ) ;
             c3_p_zfield[j] = ( 2.*kappa_y_p[j] - dt*sigma_y_p[j] ) / ( 2.*kappa_y_p[j] + dt*sigma_y_p[j] ) ;
@@ -376,7 +376,7 @@ void PML_Solver2D_Bouchard::setDomainSizeAndCoefficients( int iDim, int min_or_m
             c6_p_yfield[j] = ( 2.*kappa_y_p[j] - dt*sigma_y_p[j] ) ;
         }
 
-        for( unsigned int j = 0 ; j<ny_d ; j++ ) {
+        for( int j = 0 ; j< (int) ny_d ; j++ ) {
             c1_d_xfield[j] = ( 2.*kappa_y_d[j] - dt*sigma_y_d[j] ) / ( 2.*kappa_y_d[j] + dt*sigma_y_d[j] ) ;
             c2_d_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_d[j] + dt*sigma_y_d[j] ) ;
             c3_d_zfield[j] = ( 2.*kappa_y_d[j] - dt*sigma_y_d[j] ) / ( 2.*kappa_y_d[j] + dt*sigma_y_d[j] ) ;
