+++ src/ElectroMagnSolver/PML_Solver2D_Yee.cpp
@@ -137,7 +137,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             sigma_x_p[i] = 0. ;
         }
         // Params for other cells (PML Media) when i>=3
-        for( unsigned int i = startpml; i<nx_p ; i++ ) {
+        for( int i = startpml; i< (int) nx_p ; i++ ) {
             kappa_x_p[i] = pml_kappa_[0]->valueAt((i-startpml)*dx/length_x_pml);
             sigma_x_p[i] = pml_sigma_[0]->valueAt((i-startpml)*dx/length_x_pml);
         }
@@ -160,7 +160,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             sigma_x_d[i] = 0. ;
         }
         // Params for other cells (PML Media) when j>=4
-        for( unsigned int i = startpml+1 ; i<nx_d ; i++ ) {
+        for( int i = startpml+1 ; i< (int) nx_d ; i++ ) {
             kappa_x_d[i] = pml_kappa_[0]->valueAt((i-startpml-0.5)*dx/length_x_pml);
             sigma_x_d[i] = pml_sigma_[0]->valueAt((i-startpml-0.5)*dx/length_x_pml);
         }
@@ -195,7 +195,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             }
         }
         if (ncells_pml_max[0] != 0 ){
-            for( unsigned int i = (nx_p-1)-(ncells_pml_max[0]-1) ; i<nx_p ; i++ ) {
+            for( int i = (nx_p-1)-(ncells_pml_max[0]-1) ; i< (int) nx_p ; i++ ) {
                 kappa_x_p[i] = pml_kappa_[0]->valueAt((i - nx_p  + ncells_pml_max[0])*dx/length_x_pml_xmax);
                 sigma_x_p[i] = pml_sigma_[0]->valueAt((i - nx_p  + ncells_pml_max[0])*dx/length_x_pml_xmax);
             }
@@ -208,7 +208,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             sigma_y_p[j] = 0. ;
         }
         // Params for other cells (PML Media) when j>=3
-        for( unsigned int j = startpml ; j<ny_p ; j++ ) {
+        for( int j = startpml ; j< (int) ny_p ; j++ ) {
             kappa_y_p[j] = pml_kappa_[1]->valueAt((j-startpml)*dy/length_y_pml);
             sigma_y_p[j] = pml_sigma_[1]->valueAt((j-startpml)*dy/length_y_pml);
         }
@@ -230,7 +230,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             }
         }
         if (ncells_pml_max[0] != 0 ){
-            for( unsigned int i = (nx_p-1)-(ncells_pml_max[0]-1)+1 ; i<nx_d ; i++ ) {
+            for( int i = (nx_p-1)-(ncells_pml_max[0]-1)+1 ; i< (int) nx_d ; i++ ) {
                 kappa_x_d[i] = pml_kappa_[0]->valueAt((i - nx_p + ncells_pml_max[0] - 0.5 )*dx/length_x_pml_xmax);
                 sigma_x_d[i] = pml_sigma_[0]->valueAt((i - nx_p  + ncells_pml_max[0] - 0.5)*dx/length_x_pml_xmax);
             }
@@ -243,7 +243,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             sigma_y_d[j] = 0. ;
         }
         // Params for other cells (PML Media) when j>=4
-        for( unsigned int j = startpml+1 ; j<ny_d ; j++ ) {
+        for( int j = startpml+1 ; j< (int) ny_d ; j++ ) {
             kappa_y_d[j] = pml_kappa_[1]->valueAt((j-startpml-0.5)*dy/length_y_pml);
             sigma_y_d[j] = pml_sigma_[1]->valueAt((j-startpml-0.5)*dy/length_y_pml);
         }
@@ -255,7 +255,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
     }
 
     if ((min_or_max==0)&&(iDim==0)){
-        for( unsigned int i = 0 ; i<nx_p ; i++ ) {
+        for( int i = 0 ; i< (int) nx_p ; i++ ) {
             c1_p_zfield[i] = ( 2.*kappa_x_p[(nx_p-1)-i] - dt*sigma_x_p[(nx_p-1)-i] ) / ( 2.*kappa_x_p[(nx_p-1)-i] + dt*sigma_x_p[(nx_p-1)-i] ) ;
             c2_p_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_p[(nx_p-1)-i] + dt*sigma_x_p[(nx_p-1)-i] ) ;
             c3_p_yfield[i] = ( 2.*kappa_x_p[(nx_p-1)-i] - dt*sigma_x_p[(nx_p-1)-i] ) / ( 2.*kappa_x_p[(nx_p-1)-i] + dt*sigma_x_p[(nx_p-1)-i] ) ;
@@ -264,7 +264,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             c6_p_xfield[i] = ( 2.*kappa_x_p[(nx_p-1)-i] - dt*sigma_x_p[(nx_p-1)-i] ) ;
         }
 
-        for( unsigned int i = 0 ; i<nx_d ; i++ ) {
+        for( int i = 0 ; i< (int) nx_d ; i++ ) {
             c1_d_zfield[i] = ( 2.*kappa_x_d[(nx_d-1)-i] - dt*sigma_x_d[(nx_d-1)-i] ) / ( 2.*kappa_x_d[(nx_d-1)-i] + dt*sigma_x_d[(nx_d-1)-i] ) ;
             c2_d_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_d[(nx_d-1)-i] + dt*sigma_x_d[(nx_d-1)-i] ) ;
             c3_d_yfield[i] = ( 2.*kappa_x_d[(nx_d-1)-i] - dt*sigma_x_d[(nx_d-1)-i] ) / ( 2.*kappa_x_d[(nx_d-1)-i] + dt*sigma_x_d[(nx_d-1)-i] ) ;
@@ -274,7 +274,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
         }
     }
     else {
-        for( unsigned int i = 0 ; i<nx_p ; i++ ) {
+        for( int i = 0 ; i< (int) nx_p ; i++ ) {
             c1_p_zfield[i] = ( 2.*kappa_x_p[i] - dt*sigma_x_p[i] ) / ( 2.*kappa_x_p[i] + dt*sigma_x_p[i] ) ;
             c2_p_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_p[i] + dt*sigma_x_p[i] ) ;
             c3_p_yfield[i] = ( 2.*kappa_x_p[i] - dt*sigma_x_p[i] ) / ( 2.*kappa_x_p[i] + dt*sigma_x_p[i] ) ;
@@ -283,7 +283,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             c6_p_xfield[i] = ( 2.*kappa_x_p[i] - dt*sigma_x_p[i] ) ;
         }
 
-        for( unsigned int i = 0 ; i<nx_d ; i++ ) {
+        for( int i = 0 ; i< (int) nx_d ; i++ ) {
             c1_d_zfield[i] = ( 2.*kappa_x_d[i] - dt*sigma_x_d[i] ) / ( 2.*kappa_x_d[i] + dt*sigma_x_d[i] ) ;
             c2_d_zfield[i] = ( 2*dt ) / ( 2.*kappa_x_d[i] + dt*sigma_x_d[i] ) ;
             c3_d_yfield[i] = ( 2.*kappa_x_d[i] - dt*sigma_x_d[i] ) / ( 2.*kappa_x_d[i] + dt*sigma_x_d[i] ) ;
@@ -294,7 +294,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
     } // End X
 
     if (min_or_max==0){
-        for( unsigned int j = 0 ; j<ny_p ; j++ ) {
+        for( int j = 0 ; j< (int) ny_p ; j++ ) {
             c1_p_xfield[j] = ( 2.*kappa_y_p[(ny_p-1)-j] - dt*sigma_y_p[(ny_p-1)-j] ) / ( 2.*kappa_y_p[(ny_p-1)-j] + dt*sigma_y_p[(ny_p-1)-j] ) ;
             c2_p_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_p[(ny_p-1)-j] + dt*sigma_y_p[(ny_p-1)-j] ) ;
             c3_p_zfield[j] = ( 2.*kappa_y_p[(ny_p-1)-j] - dt*sigma_y_p[(ny_p-1)-j] ) / ( 2.*kappa_y_p[(ny_p-1)-j] + dt*sigma_y_p[(ny_p-1)-j] ) ;
@@ -303,7 +303,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             c6_p_yfield[j] = ( 2.*kappa_y_p[(ny_p-1)-j] - dt*sigma_y_p[(ny_p-1)-j] ) ;
         }
 
-        for( unsigned int j = 0 ; j<ny_d ; j++ ) {
+        for( int j = 0 ; j< (int) ny_d ; j++ ) {
             c1_d_xfield[j] = ( 2.*kappa_y_d[(ny_d-1)-j] - dt*sigma_y_d[(ny_d-1)-j] ) / ( 2.*kappa_y_d[(ny_d-1)-j] + dt*sigma_y_d[(ny_d-1)-j] ) ;
             c2_d_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_d[(ny_d-1)-j] + dt*sigma_y_d[(ny_d-1)-j] ) ;
             c3_d_zfield[j] = ( 2.*kappa_y_d[(ny_d-1)-j] - dt*sigma_y_d[(ny_d-1)-j] ) / ( 2.*kappa_y_d[(ny_d-1)-j] + dt*sigma_y_d[(ny_d-1)-j] ) ;
@@ -313,7 +313,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
         }
     }
     else if (min_or_max==1){
-        for( unsigned int j = 0 ; j<ny_p ; j++ ) {
+        for( int j = 0 ; j< (int) ny_p ; j++ ) {
             c1_p_xfield[j] = ( 2.*kappa_y_p[j] - dt*sigma_y_p[j] ) / ( 2.*kappa_y_p[j] + dt*sigma_y_p[j] ) ;
             c2_p_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_p[j] + dt*sigma_y_p[j] ) ;
             c3_p_zfield[j] = ( 2.*kappa_y_p[j] - dt*sigma_y_p[j] ) / ( 2.*kappa_y_p[j] + dt*sigma_y_p[j] ) ;
@@ -322,7 +322,7 @@ void PML_Solver2D_Yee::setDomainSizeAndCoefficients( int iDim, int min_or_max, s
             c6_p_yfield[j] = ( 2.*kappa_y_p[j] - dt*sigma_y_p[j] ) ;
         }
 
-        for( unsigned int j = 0 ; j<ny_d ; j++ ) {
+        for( int j = 0 ; j< (int) ny_d ; j++ ) {
             c1_d_xfield[j] = ( 2.*kappa_y_d[j] - dt*sigma_y_d[j] ) / ( 2.*kappa_y_d[j] + dt*sigma_y_d[j] ) ;
             c2_d_xfield[j] = ( 2*dt ) / ( 2.*kappa_y_d[j] + dt*sigma_y_d[j] ) ;
             c3_d_zfield[j] = ( 2.*kappa_y_d[j] - dt*sigma_y_d[j] ) / ( 2.*kappa_y_d[j] + dt*sigma_y_d[j] ) ;
